# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "server API",
    "version": "1.0.0",
    "description": "Machine-readable API contract for the Express backend in server."
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      },
      "apiKeyAuthScheme": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Alternative format: `Authorization: ApiKey <apiKey>`"
      },
      "inboundBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token for worker ingress"
      }
    },
    "schemas": {
      "RequestId": {
        "type": "string"
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "requestId"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {},
          "requestId": {
            "$ref": "#/components/schemas/RequestId"
          }
        }
      },
...
      "SimpleSuccessResponse": {
        "type": "object",
        "required": [
          "success",
          "requestId"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "requestId": {
            "$ref": "#/components/schemas/RequestId"
          }
        }
      },
...
      "BlockedSenderCreatedBy": {
        "type": "object",
        "nullable": true,
        "properties": {
          "userId": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        }
      },
      "BlockedSender": {
        "type": "object",
        "required": [
          "id",
          "patternType",
          "pattern",
          "scope",
          "domain",
          "reason",
          "status",
          "matchCount",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "patternType": {
            "type": "string",
            "enum": [
              "email",
              "domain"
            ],
            "description": "domain also blocks every subdomain of the pattern"
          },
          "pattern": {
            "type": "string"
          },
          "scope": {
            "type": "string",
            "enum": [
              "global",
              "domain"
            ],
            "description": "domain scope only blocks mail arriving at the linked recipient domain"
          },
          "domain": {
            "type": "string",
            "nullable": true,
            "description": "Recipient domain the rule is restricted to, null for global rules"
          },
          "reason": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled"
            ]
          },
          "matchCount": {
            "type": "integer"
          },
          "lastMatchedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdBy": {
            "$ref": "#/components/schemas/BlockedSenderCreatedBy"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BlockedSenderCreateRequest": {
        "type": "object",
        "required": [
          "pattern"
        ],
        "properties": {
          "pattern": {
            "type": "string",
            "description": "Sender address (user@example.com) or sender domain (example.com)"
          },
          "patternType": {
            "type": "string",
            "enum": [
              "email",
              "domain"
            ],
            "description": "Inferred from pattern when omitted"
          },
          "domain": {
            "type": "string",
            "nullable": true,
            "description": "Restrict the rule to a single recipient domain"
          },
          "reason": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled"
            ]
          }
        }
      },
      "BlockedSenderUpdateRequest": {
        "type": "object",
        "properties": {
          "pattern": {
            "type": "string"
          },
          "patternType": {
            "type": "string",
            "enum": [
              "email",
              "domain"
            ]
          },
          "domain": {
            "type": "string",
            "nullable": true
          },
          "reason": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled"
            ]
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/v1/blocked-senders": {
      "get": {
        "tags": [
          "BlockedSenders"
        ],
        "summary": "List blocked senders",
        "description": "Global admin only",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "apiKeyAuthScheme": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "patternType",
            "schema": {
              "type": "string",
              "enum": [
                "email",
                "domain"
              ]
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "disabled"
              ]
            }
          },
          {
            "in": "query",
            "name": "scope",
            "schema": {
              "type": "string",
              "enum": [
                "global",
                "domain"
              ]
            }
          },
          {
            "in": "query",
            "name": "domain",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blocked sender list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "blockedSenders": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BlockedSender"
                      }
                    },
                    "requestId": {
                      "$ref": "#/components/schemas/RequestId"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "BlockedSenders"
        ],
        "summary": "Block a sender",
        "description": "Global admin only",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "apiKeyAuthScheme": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockedSenderCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Blocked sender created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "blockedSender": {
                      "$ref": "#/components/schemas/BlockedSender"
                    },
                    "requestId": {
                      "$ref": "#/components/schemas/RequestId"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Sender already blocked for the requested scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/blocked-senders/{blockedSenderId}": {
      "get": {
        "tags": [
          "BlockedSenders"
        ],
        "summary": "Get blocked sender by ID",
        "description": "Global admin only",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "apiKeyAuthScheme": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "blockedSenderId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blocked sender details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "blockedSender": {
                      "$ref": "#/components/schemas/BlockedSender"
                    },
                    "requestId": {
                      "$ref": "#/components/schemas/RequestId"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "BlockedSenders"
        ],
        "summary": "Update blocked sender",
        "description": "Global admin only",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "apiKeyAuthScheme": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "blockedSenderId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockedSenderUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Blocked sender updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "blockedSender": {
                      "$ref": "#/components/schemas/BlockedSender"
                    },
                    "requestId": {
                      "$ref": "#/components/schemas/RequestId"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "BlockedSenders"
        ],
        "summary": "Delete blocked sender",
        "description": "Global admin only",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "apiKeyAuthScheme": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "blockedSenderId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blocked sender deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleSuccessResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}